add booking validation methods into a concern
