+++ app/models/booking.rb
@@ -25,6 +25,8 @@
 #  fk_rails_...  (service_id => services.id)
 #
 class Booking < ApplicationRecord
+  include BookingValidation
+
   belongs_to :client, class_name: "User"
 
   belongs_to :business
@@ -64,31 +66,6 @@ class Booking < ApplicationRecord
     
   private
 
-  # Ensure the booking fits within the business's operating hours
-  def within_business_hours
-    return unless business && started_at && ended_at
-
-    # Get the day of the week for the booking
-    day_of_week = started_at.in_time_zone(business.time_zone).strftime("%A")
-    
-    # Find the business hours for that day
-    business_hour = business.business_hours.find_by(day_of_the_week: day_of_week)
-    
-    if !business_hour.closed
-      opening_time = business_hour.adjusted_opening_time.strftime("%H:%M:%S")
-      closing_time = business_hour.adjusted_closing_time.strftime("%H:%M:%S")
-  
-    # Validate against the business's open and close times
-      if business_hour.closed || business_hour.nil?
-        errors.add(:base, "The business is closed on #{day_of_week}.")
-      elsif started_at.strftime("%H:%M:%S") < opening_time || ended_at.strftime("%H:%M:%S") > closing_time
-        errors.add(:base, "The booking time is outside the business hours for #{day_of_week}.")
-      end
-    else
-      errors.add(:base, "The business is closed on #{day_of_week}.")
-    end
-  end
-
   def ensure_ended_at_has_value
     if !ended_at
       self.ended_at = started_at + service.duration.minutes
@@ -101,29 +78,6 @@ class Booking < ApplicationRecord
     tz = ActiveSupport::TimeZone[business.time_zone]
     started_at = tz.parse(started_at.to_s) if started_at.present?
     ended_at = tz.parse(ended_at.to_s) if ended_at.present?
-  end
-
-    # Validation logic for booking time
-    def validate_booking_time(business, proposed_started_at, proposed_ended_at)
-      business_bookings = business.accepted_received_bookings
-      client_bookings = client.accepted_sent_bookings
-  
-      bookings = business_bookings + client_bookings
-      
-      bookings.none? do |booking|
-        (proposed_started_at < booking.ended_at) && (proposed_ended_at > booking.started_at)
-      end
-    end
-
-  def time_slot_availability
-    business = self.business
-
-    # Ensure we have a business and a valid time range
-    if business && started_at && ended_at
-      unless validate_booking_time(business, started_at, ended_at)
-        errors.add(:base, "The selected time slot is unavailable.")
-      end
-    end
   end
 end
